requirements.txt
flask
google-cloud-datastore
google-cloud-bigquery
google-cloud-ndb
PyYAML
pandas
numpy
Flask-Cors
google-auth
requests
google-cloud-storage
beautifulsoup4
Pillow
mgrs
#Werkzeug==0.16.0
webdriver_manager
flask-restplus
configparser